Move the AppLayout from the Widget to the Abstract folder
